Adding basic_auth feature and adding --non-interactive to all svn commands.
The basic_auth feature allows the manifest to specify an optional
basic_auth_username and basic_auth_password to be sent to the subversion
server over HTTP(S).

Adding --non-interactive to all svn commands just lets svn know that
there's not a user on the other end to send y/n questions.